feat(wasmcloud-host): enable preliminary support for multiple lattices on a single host
This implements initial support for multiple lattices running on a
single host by allowing the `--lattice` flag to be specified multiple
times. The host code itself is also shuffled around to support this
functionality, since each lattice now needs to handle its own
subscriptions, cleanup signals, etc. As a result, the host code is now
much more compact than it was before.

This work is the first step towards implementing full multitenancy
support in the host as outlined in #3120. The next step is to add an
experimental flag to enable lattices to be defined in the control plane
and have the host dynamically listen to requests to listen on lattices
as they are provisioned.
