+++ crates/host/src/wasmbus/host_config.rs
@@ -33,8 +33,9 @@ pub struct Host {
     pub rpc_key: Option<Arc<KeyPair>>,
     /// Whether to require TLS for RPC connection
     pub rpc_tls: bool,
-    /// The lattice the host belongs to
-    pub lattice: Arc<str>,
+    /// The lattices the host belongs to
+    /// TODO: should this actually be Vec<Arc<str>>?
+    pub lattices: Vec<String>,
     /// The domain to use for host Jetstream operations
     pub js_domain: Option<String>,
     /// Labels (key-value pairs) to add to the host
@@ -100,7 +101,7 @@ impl Default for Host {
             rpc_jwt: None,
             rpc_key: None,
             rpc_tls: false,
-            lattice: "default".into(),
+            lattices: vec!["default".to_string()],
             js_domain: None,
             labels: HashMap::default(),
             host_key: None,
