+++ tests/interfaces.rs
@@ -78,7 +78,7 @@ async fn interfaces() -> anyhow::Result<()> {
         .lattice(LATTICE.to_string())
         .build();
     // Build the host
-    let host = WasmCloudTestHost::start(&nats_url, LATTICE)
+    let host = WasmCloudTestHost::start(&nats_url, vec![LATTICE.to_string()])
         .await
         .context("failed to start test host")?;
